split examples by endpoint
